+++ librtl/src/lib_add_times.c
@@ -1,38 +1,38 @@
 /*
 
-       VMSTIME.C  v1.1
+ VMSTIME.C  v1.1
 
-       Author: Paul Nankervis
+ Author: Paul Nankervis
 
-       Please send bug reports or requests for enhancement
-       or improvement via email to:     PaulNank@au1.ibm.com
+ Please send bug reports or requests for enhancement
+ or improvement via email to:     PaulNank@au1.ibm.com
 
 
-       This module contains versions of the VMS time routines
-       sys$numtim(), sys$asctim() and friends... They are
-       intended to be compatible with the routines of the same
-       name on a VMS system (so descriptors feature regularly!)
+ This module contains versions of the VMS time routines
+ sys$numtim(), sys$asctim() and friends... They are
+ intended to be compatible with the routines of the same
+ name on a VMS system (so descriptors feature regularly!)
 
-       This code relies on being able to manipluate day numbers
-       and times using 32 bit arithmetic to crack a VMS quadword
-       byte by byte. If your C compiler doesn't have 32 bit int
-       fields give up now! On a 64 bit systems this code could
-       be modified to do 64 bit operations directly....
+ This code relies on being able to manipluate day numbers
+ and times using 32 bit arithmetic to crack a VMS quadword
+ byte by byte. If your C compiler doesn't have 32 bit int
+ fields give up now! On a 64 bit systems this code could
+ be modified to do 64 bit operations directly....
 
-       One advantage of doing arihmetic byte by byte is that
-       the code does not depend on what 'endian' the target
-       machine is - it will always treat bytes in the same order!
-       (Hopefully VMS time bytes will always be in the same order!)
+ One advantage of doing arihmetic byte by byte is that
+ the code does not depend on what 'endian' the target
+ machine is - it will always treat bytes in the same order!
+ (Hopefully VMS time bytes will always be in the same order!)
 
-       A couple of stupid questions to go on with:-
-           o OK, I give up! What is the difference between a zero
-             date and a zero delta time?
-           o Anyone notice that the use of 16 bit words in
-             sys$numtim restricts delta times to 65535 days?
+ A couple of stupid questions to go on with:-
+ o OK, I give up! What is the difference between a zero
+ date and a zero delta time?
+ o Anyone notice that the use of 16 bit words in
+ sys$numtim restricts delta times to 65535 days?
 
-                                       Paul Nankervis
+ Paul Nankervis
 
-*/
+ */
 
 #include <stddef.h>
 #include "libdef.h"
@@ -43,9 +43,7 @@ struct TIME
     unsigned char time[8];
 };
 
-
-unsigned long lib$add_times(const void *time1a, const void *time2a,
-                            void *result)
+int lib$add_times(const void *time1a, const void *time2a, void *result)
 {
     const struct TIME* time1 = (const struct TIME*) time1a;
     const struct TIME* time2 = (const struct TIME*) time2a;
@@ -54,18 +52,18 @@ unsigned long lib$add_times(const void *time1a, const void *time2a,
     {
         if (time2->time[7] & 0x80)
         {
-            return lib$addx(time1,time2,result,NULL);
+            return lib$addx(time1, time2, result, NULL);
         }
         else
         {
-            return lib$subx(time2,time1,result,NULL);
+            return lib$subx(time2, time1, result, NULL);
         }
     }
     else
     {
         if (time2->time[7] & 0x80)
         {
-            return lib$subx(time1,time2,result,NULL);
+            return lib$subx(time1, time2, result, NULL);
         }
         else
         {
