+++ librtl/src/str_element.c
@@ -69,28 +69,26 @@
  * str$element
  *
  */
-unsigned long str$element(struct dsc$descriptor_s* destination_string,
-                          const long* find_element_number,
-                          const struct dsc$descriptor_s* delimiter_string,
-                          const struct dsc$descriptor_s* source_string)
+int str$element(struct dsc$descriptor_s* destination_string, const long* find_element_number,
+        const struct dsc$descriptor_s* delimiter_string, const struct dsc$descriptor_s* source_string)
 {
-    char* s1_ptr;           /*Pointer to first string */
-    unsigned short s1_length;   /* Length of first string */
-    char* s2_ptr;           /* Pointer to second string */
-    unsigned short s2_length;   /* Length of second string */
-    char* s3_ptr    ;       /* Pointer to third string */
-    unsigned short s3_length;   /* Length of third string */
-    short res_ptr;          /* Index into source*/
-    unsigned short res_length;  /* Index to end  of element */
-    char delimiter;         /* Delimiter character */
-    unsigned long result1;
-    unsigned long result = STR$_NORMAL; /* Result */
+    char* s1_ptr; /*Pointer to first string */
+    unsigned short s1_length; /* Length of first string */
+    char* s2_ptr; /* Pointer to second string */
+    unsigned short s2_length; /* Length of second string */
+    char* s3_ptr; /* Pointer to third string */
+    unsigned short s3_length; /* Length of third string */
+    short res_ptr; /* Index into source*/
+    unsigned short res_length; /* Index to end  of element */
+    char delimiter; /* Delimiter character */
+    int result1;
+    int result = STR$_NORMAL; /* Result */
     int element_number;
 
     element_number = (int) *find_element_number;
 
 //  Did we get a numeric value
-    if ( isdigit((char) element_number + '0') == 0 )
+    if ( isdigit((char) element_number + '0') == 0)
         return STR$_ILLSTRSPE;
 
     /*
@@ -101,7 +99,6 @@ unsigned long str$element(struct dsc$descriptor_s* destination_string,
         element_number = 0;
     }
 
-
     /*
      * Analyze strings
      */
@@ -110,9 +107,9 @@ unsigned long str$element(struct dsc$descriptor_s* destination_string,
     str$analyze_sdesc(delimiter_string, &s3_length, &s3_ptr);
 
 //  If no delimiter's where found return entire string
-    if ( (s2_ptr != NULL) && (s3_ptr != NULL) )
+    if ((s2_ptr != NULL) && (s3_ptr != NULL))
     {
-        if ( strstr (s2_ptr, s3_ptr) == NULL)
+        if (strstr(s2_ptr, s3_ptr) == NULL)
         {
 //          str$copy_dx (destination_string,source_string);
             return SS$_NORMAL;
@@ -154,8 +151,7 @@ unsigned long str$element(struct dsc$descriptor_s* destination_string,
         /*
          * Look for ending delimiter
          */
-        while (((res_ptr + res_length) < s2_length) &&
-                (s2_ptr[res_ptr + res_length] != delimiter))
+        while (((res_ptr + res_length) < s2_length) && (s2_ptr[res_ptr + res_length] != delimiter))
         {
             res_length++;
         }
@@ -164,18 +160,18 @@ unsigned long str$element(struct dsc$descriptor_s* destination_string,
     /*
      * Copy over the result
      */
-    if ( res_ptr < s2_length )
+    if (res_ptr < s2_length)
     {
         s1_ptr = s2_ptr + res_ptr;
         result1 = str$copy_r(destination_string, &res_length, s1_ptr);
     }
     /*
-        else
-        {   str$free1_dx (destination_string);
-            result1 = STR$_NOELEM;
-            result  = STR$_NOELEM;
-        }
-    */
+     else
+     {   str$free1_dx (destination_string);
+     result1 = STR$_NOELEM;
+     result  = STR$_NOELEM;
+     }
+     */
     if (result1 == STR$_NORMAL)
     {
         result = SS$_NORMAL;
