+++ librtl/src/str_find_first_in_set.c
@@ -37,16 +37,14 @@
  *      in the source string where the first match is found.
  *
  */
-long str$find_first_in_set(
-    const struct dsc$descriptor_s* first_source_string,
-    const struct dsc$descriptor_s* second_source_string)
+int str$find_first_in_set(const struct dsc$descriptor_s* first_source_string, const struct dsc$descriptor_s* second_source_string)
 {
-    char* s1_ptr;           /* Pointer to first string */
-    unsigned short s1_length;   /* Length of first string */
-    char* s2_ptr;           /* Pointer to second string */
-    unsigned short s2_length;   /* Length of second string */
-    int loop1;          /* Outer loop */
-    int loop2;          /* Inner loop */
+    char* s1_ptr; /* Pointer to first string */
+    unsigned short s1_length; /* Length of first string */
+    char* s2_ptr; /* Pointer to second string */
+    unsigned short s2_length; /* Length of second string */
+    int loop1; /* Outer loop */
+    int loop2; /* Inner loop */
 
     /*
      * Analyze source strings
@@ -97,5 +95,4 @@ long str$find_first_in_set(
      */
     return 0;
 }
-
